Revert "Ensure thatPreferredCursorExecution is the only configuration flag to control cursor preference."
This reverts commit 07cb863d257909a7ea961c00c5a281f174200daf.
